fix: Add generic evaluator function
