+++ py/torch_tensorrt/dynamo/conversion/impl/__init__.py
@@ -6,7 +6,6 @@ from . import (
     condition,
     elementwise,
     embedding,
-    evaluators,
     matmul,
     normalization,
     permutation,
